backend/requirements.txt
fastapi \
uvicorn \
pymupdf \
openai \
pinecone-client \
python-docx \
nltk